Move parsing to parse.go
